Moved pupilParser to other Project and did refactor

+++ iTechArt.Repository/BusinessModels/Pupil.cs
@@ -8,63 +8,63 @@ namespace iTechArt.Repository.BusinessModels
     internal sealed class Pupil : IPupil
     {
         /// <summary>
-        /// Gets or sets id of pupil
+        /// Gets or sets id of pupil.
         /// </summary>
         public long Id { get; internal set; }
 
         /// <summary>
-        /// Gets or sets the name of pupil
+        /// Gets or sets the name of pupil.
         /// </summary>
         public string FirstName { get; internal set; }
 
         /// <summary>
-        /// Gets or sets surname of pupil
+        /// Gets or sets surname of pupil.
         /// </summary>
         public string LastName { get; internal set; }
 
         /// <summary>
-        /// Gets or sets birthdate of pupil
+        /// Gets or sets birthdate of pupil.
         /// </summary>
         [JsonConverter(typeof(DateOnlyJsonConverter))]
         public DateOnly DateOfBirth { get; internal set; }
 
         /// <summary>
-        /// Gets or sets gender of pupil
+        /// Gets or sets gender of pupil.
         /// </summary>
         public Gender Gender { get; internal set; }
 
         /// <summary>
-        /// Gets or sets phone number of pupil
+        /// Gets or sets phone number of pupil.
         /// </summary>
         public string PhoneNumber { get; internal set; }
 
         /// <summary>
-        /// Gets or sets address of pupil
+        /// Gets or sets address of pupil.
         /// </summary>
         public string Address { get; internal set; }
 
         /// <summary>
-        /// Gets or sets the city, where pupil lives
+        /// Gets or sets the city, where pupil lives.
         /// </summary>
         public string City { get; internal set; }
 
         /// <summary>
-        /// Gets or sets the name of school, where pupil study
+        /// Gets or sets the name of school, where pupil study.
         /// </summary>
         public string SchoolName { get; internal set; }
 
         /// <summary>
-        /// Gets or sets the grade of study of pupil
+        /// Gets or sets the grade of study of pupil.
         /// </summary>
-        public byte Grade { get; internal set; }
+        public int Grade { get; internal set; }
 
         /// <summary>
-        /// Gets or sets the language of education
+        /// Gets or sets the language of education.
         /// </summary>
         public CourseLanguage CourseLanguage { get; internal set; }
 
         /// <summary>
-        /// Gets or sets shift of study
+        /// Gets or sets shift of study.
         /// </summary>
         public Shift Shift { get; internal set; }
     }
